fix: int64_t loop in coef_stats + atomic realloc in model_stub
- coef_stats.c: int→int64_t loop var for mat->nnz scan (>2^31 safety)
- model_stub.c: realloc all 5 arrays into temps, check all for NULL,
  then commit atomically. Prevents partial-growth inconsistency.

Closes convexfeld-vlja, convexfeld-yhmx. 40/40 tests pass.

+++ src/analysis/coef_stats.c
@@ -77,7 +77,7 @@ int cxf_compute_coef_stats(CxfModel *model,
     /* Scan matrix coefficients (CSC format) */
     if (model->matrix != NULL && model->matrix->values != NULL) {
         MatrixData *mat = model->matrix;
-        for (int k = 0; k < mat->nnz; k++) {
+        for (int64_t k = 0; k < mat->nnz; k++) {
             double val = fabs(mat->values[k]);
             if (val > 0.0) {
                 if (val < m_min) m_min = val;
